Fixed per-epoch shuffle, updated shuffle methods for DataLoadWrapper
Prior to this update, there was no reordering of subset presentation.  Though subsets themselves were shuffled, the overall order of presentation stayed the same.
DataLoadWrapper now contains two shuffle methods: one for subset at getitem time, and one for post epoch reorder. Generation of subsets is now more simple as well.  Train-validation split is in prototype, but the issue of validation for this specific problem remains elusive. As such, the train.py module has been added to, with functions to handle load of previous data or to generate new data.